eduai-ai-service/pyproject.toml
[tool.poetry]
name = "eduai-ai-service"
version = "1.0.0"
description = "AI-powered course generation microservice for EduAI platform"
authors = ["EduAI Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
openai = "^1.10.0"
anthropic = "^0.18.0"
google-generativeai = "^0.3.2"
pdfplumber = "^0.10.3"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
